Check the object is a valid region series

A region series is a terra::SpatRasterDataset for which each sub-dataset is a variable, and all variables have the same number of time steps.

Usage

is_region_series(x, strict = FALSE)

Arguments

x

a terra::SpatRasterDataset representing a time series of regional reconstructions obtained from region_series.

strict

a boolean defining whether to preform a thorough test (see description above for details).

Value

TRUE if the object is a region series

Details

The standard test only checks that each SpatRaster has the same number of layers. The more thorough test (obtained with strict=TRUE) actually checks that all time steps are identical by comparing the result of terra::time applied to each variable