pastclim: Manipulate Time Series of Palaeoclimate Reconstructions

Methods to easily extract and manipulate palaeoclimate reconstructions for ecological and anthropological analyses, as described in Leonardi et al. (2023) doi:10.1111/ecog.06481 .

See also

Author

Maintainer: Andrea Manica am315@cam.ac.uk

Authors:

Michela Leonardi

Other contributors:

Emily Y. Hallet [contributor]

Robert Beyer [contributor]

Mario Krapp [contributor]

Andrea V. Pozzi [contributor]