Extract a climate slice for a region

This function extracts a slice of one or more climate variables for a given dataset covering a region (or the whole world). The function returns a SpatRaster terra::SpatRaster object, with each variable as a layer.

Usage

region_slice(
  time_bp,
  bio_variables,
  dataset,
  path_to_nc = NULL,
  ext = NULL,
  crop = NULL
)

Arguments

time_bp

the time slice in years before present (negative values represent time before present, positive values time in the future). The slice needs to exist in the dataset. To check which slices are available, you can use get_time_steps.

bio_variables

vector of names of variables to be extracted

dataset

string defining the dataset to use. If set to "custom", then a single nc file is used from "path_to_nc"

path_to_nc

the path to the custom nc file containing the palaeoclimate reconstructions. All the variables of interest need to be included in this file.

ext

an extent, coded as numeric vector (length=4; order= xmin, xmax, ymin, ymax) or a terra::SpatExtent object. If NULL, the full extent of the reconstruction is given.

crop

a polygon used to crop the reconstructions (e.g. the outline of a continental mass). A sf:sfg or a terra::SpatVector object is used to define the polygon.

Value

a SpatRaster terra::SpatRaster object, with each variable as a layer.