Documentation for the Krapp2021 dataset

This dataset covers the last 800k years, at intervals of 1k years, and a resolution of 0.5 degrees in latitude and longitude.

Details

If you use this dataset, make sure to cite the original publication:

Krapp, M., Beyer, R.M., Edmundson, S.L. et al. A statistics-based reconstruction of high-resolution global terrestrial climate for the last 800,000 years. Sci Data 8, 228 (2021). doi:doi.org/10.1038/s41597-021-01009-3

The version included in `pastclim` has the ice sheets masked.

Note that, for bio15, we use the corrected version, which follows https://pubs.usgs.gov/ds/691/ds691.pdf

Changelog

v1.1.0 Added monthly variables. Files can be downloaded from: https://zenodo.org/record/7065055

v1.0.0 Remove ice sheets and use correct formula for bio15. Files can be downloaded from: doi:doi.org/10.6084/m9.figshare.19733680.v1